+++ packages/llm/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @customize-agent/llm
 
+## 2.0.0
+
+### Major Changes
+
+- 6744afe: 优化、拓展、兼容、新增
+
+### Patch Changes
+
+- Updated dependencies [6744afe]
+  - @customize-agent/types@2.0.0
+
 ## 1.0.2
 
 ### Patch Changes