+++ packages/memory/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @customize-agent/memory
 
+## 2.0.0
+
+### Major Changes
+
+- 6744afe: 优化、拓展、兼容、新增
+
 ## 1.0.2
 
 ### Patch Changes